gitserver: Move cloneurl into internal package
Was a merge race of introducing this package and introducing the internal package.
